# server/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Shawarma Server API",
        "title": "Shawarma API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:4000",
    "basePath": "/",
    "paths": {
        "/v1/register/user": {
            "post": {
                "description": "Register a new user with the provided user details",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "Register a new user",
                "parameters": [
                    {
                        "description": "User details",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                ],
                "responses": {
                    "202": {
                        "description": "Accepted",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/v1/repos": {
            "get": {
                "description": "Retrieve all users' repositories",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "git"
                ],
                "summary": "Get all repositories",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/github_com_zerodoctor_shawarma_pkg_model.Repository"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/v1/user": {
            "get": {
                "description": "Retrieve user information by state and remote type",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "Get user by state",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Remote type",
                        "name": "type",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Remote state",
                        "name": "state",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/v1/user/{name}": {
            "get": {
                "description": "Retrieve user information by username",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "Get user by name",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Username",
                        "name": "name",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_zerodoctor_shawarma_pkg_model.User"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "github_com_zerodoctor_shawarma_pkg_model.Branch": {
            "type": "object",
            "properties": {
                "commits": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_zerodoctor_shawarma_pkg_model.Commit"
                    }
                },
                "created_at": {
                    "type": "string"
                },
                "hash": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "moified_at": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "repo_id": {
                    "type": "string"
                }
            }
        },
        "github_com_zerodoctor_shawarma_pkg_model.Commit": {
            "type": "object",
            "properties": {
                "author": {
                    "type": "string"
                },
                "commit": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_zerodoctor_shawarma_pkg_model.Commit"
                    }
                },
                "created_at": {
                    "type": "string"
                },
                "hash": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "github_com_zerodoctor_shawarma_pkg_model.Environment": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "data": {
                    "type": "string"
                },
                "key": {
                    "type": "string"
                },
                "modified_at": {
                    "type": "string"
                },
                "org_id": {
                    "type": "string"
                },
                "protected": {
                    "type": "boolean"
                },
                "repo_id": {
                    "type": "string"
                }
            }
        },
        "github_com_zerodoctor_shawarma_pkg_model.Organization": {
            "type": "object",
            "properties": {
                "avatar_url": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "env": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "environments": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_zerodoctor_shawarma_pkg_model.Environment"
                    }
                },
                "modifiedAt": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "owner": {
                    "type": "string"
                },
                "repository": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_zerodoctor_shawarma_pkg_model.Repository"
                    }
                },
                "uuid": {
                    "type": "string"
                }
            }
        },
        "github_com_zerodoctor_shawarma_pkg_model.Repository": {
            "type": "object",
            "properties": {
                "active": {
                    "$ref": "#/definitions/sql.NullBool"
                },
                "branches": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_zerodoctor_shawarma_pkg_model.Branch"
                    }
                },
                "created_at": {
                    "type": "string"
                },
                "default_branch": {
                    "type": "string"
                },
                "env": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "environments": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_zerodoctor_shawarma_pkg_model.Environment"
                    }
                },
                "modified_at": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "owner": {
                    "type": "string"
                },
                "owner_id": {
                    "type": "string"
                },
                "owner_type": {
                    "type": "string"
                },
                "uuid": {
                    "type": "string"
                }
            }
        },
        "github_com_zerodoctor_shawarma_pkg_model.User": {
            "type": "object",
            "properties": {
                "avatar_url": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "git_remote": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "is_owner": {
                    "type": "boolean"
                },
                "modified_at": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "organizations": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_zerodoctor_shawarma_pkg_model.Organization"
                    }
                },
                "polls": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "repositories": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_zerodoctor_shawarma_pkg_model.Repository"
                    }
                },
                "session": {
                    "type": "string"
                },
                "tokens": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "uuid": {
                    "type": "string"
                }
            }
        },
        "sql.NullBool": {
            "type": "object",
            "properties": {
                "bool": {
                    "type": "boolean"
                },
                "valid": {
                    "description": "Valid is true if Bool is not NULL",
                    "type": "boolean"
                }
            }
        }
    }
}